requirements.txt
PyYAML==5.3.1
python-slugify==4.0.0
fabric==2.5.0
